+++ src/android/OSInAppBrowser.kt
@@ -2,9 +2,12 @@ package com.outsystems.plugins.inappbrowser.osinappbrowser
 
 import com.google.android.gms.common.ConnectionResult
 import com.google.android.gms.common.GoogleApiAvailability
+import com.google.gson.Gson
 import com.outsystems.plugins.inappbrowser.osinappbrowserlib.OSIABEngine
-import com.outsystems.plugins.inappbrowser.osinappbrowserlib.routeradapters.ApplicationContextAdapter
-import com.outsystems.plugins.inappbrowser.osinappbrowserlib.routeradapters.OSIABApplicationRouterAdapter
+import com.outsystems.plugins.inappbrowser.osinappbrowserlib.models.OSIABToolbarPosition
+import com.outsystems.plugins.inappbrowser.osinappbrowserlib.models.OSIABWebViewOptions
+import com.outsystems.plugins.inappbrowser.osinappbrowserlib.routeradapters.OSIABExternalBrowserRouterAdapter
+import com.outsystems.plugins.inappbrowser.osinappbrowserlib.routeradapters.OSIABWebViewRouterAdapter
 import com.outsystems.plugins.oscordova.CordovaImplementation
 import org.apache.cordova.CallbackContext
 import org.apache.cordova.CordovaInterface
@@ -13,14 +16,14 @@ import org.json.JSONArray
 
 class OSInAppBrowser: CordovaImplementation() {
     override var callbackContext: CallbackContext? = null
-
     private var engine: OSIABEngine? = null
+    private val gson by lazy { Gson() }
 
     override fun initialize(cordova: CordovaInterface, webView: CordovaWebView) {
         super.initialize(cordova, webView)
-        val applicationDelegate = ApplicationContextAdapter(cordova.context)
-        val router = OSIABApplicationRouterAdapter(applicationDelegate)
-        this.engine = OSIABEngine(router)
+        val externalBrowserRouter = OSIABExternalBrowserRouterAdapter(cordova.context)
+        val webViewRouter = OSIABWebViewRouterAdapter(cordova.context)
+        this.engine = OSIABEngine(externalBrowserRouter, webViewRouter)
     }
 
     override fun execute(
@@ -34,11 +37,18 @@ class OSInAppBrowser: CordovaImplementation() {
             "openInExternalBrowser" -> {
                 openInExternalBrowser(args)
             }
+            "openInWebView" -> {
+                openInWebView(args)
+            }
         }
 
         return true
     }
 
+    /**
+     * Calls the openExternalBrowser method of OSIABEngine to open the url in the device's browser app
+     * @param args JSONArray that contains the parameters to parse (e.g. url to open)
+     */
     private fun openInExternalBrowser(args: JSONArray) {
         try {
             val argumentsDictionary = args.getJSONObject(0)
@@ -57,6 +67,32 @@ class OSInAppBrowser: CordovaImplementation() {
         }
         catch (e: Exception) {
             sendPluginResult(null, OSInAppBrowserError.INPUT_ARGUMENTS_ISSUE.toPair())
+        }
+    }
+
+    /**
+     * Calls the openWebView method of OSIABEngine to open the url in a WebView
+     * @param args JSONArray that contains the parameters to parse (e.g. url to open)
+     */
+    private fun openInWebView(args: JSONArray) {
+        try {
+            val arguments = args.getJSONObject(0)
+            val url = arguments.getString("url")
+            val webViewOptions = buildWebViewOptions(arguments.getString("options"))
+
+            engine?.openWebView(url, webViewOptions) { success ->
+                if (success) {
+                    sendPluginResult("success", null)
+                } else {
+                    sendPluginResult(
+                        null,
+                        OSInAppBrowserError.OPEN_WEB_VIEW_FAILED.toPair()
+                    )
+                }
+            }
+        }
+        catch (e: Exception) {
+            sendPluginResult(null, OSInAppBrowserError.INPUT_ARGUMENTS_WEB_VIEW_ISSUE.toPair())
         }
     }
 
@@ -80,4 +116,29 @@ class OSInAppBrowser: CordovaImplementation() {
         }
         return true
     }
+
+    /**
+     * Parses options that come in JSON to a 'OSInAppBrowserWebViewInputArguments'.
+     * Then, it uses the newly created object to create a 'OSIABWebViewOptions' object.
+     * @param options The options to open the URL in a WebView, in a JSON string.
+     */
+    private fun buildWebViewOptions(options: String): OSIABWebViewOptions {
+        return gson.fromJson(options, OSInAppBrowserWebViewInputArguments::class.java).let {
+            OSIABWebViewOptions(
+                it.showURL ?: true,
+                it.showToolbar ?: true,
+                it.clearCache ?: true,
+                it.clearSessionCache ?: true,
+                it.mediaPlaybackRequiresUserAction ?: false,
+                it.closeButtonText ?: "Close",
+                it.toolbarPosition ?: OSIABToolbarPosition.TOP,
+                it.leftToRight ?: false,
+                it.showNavigationButtons ?: true,
+                it.android.allowZoom ?: true,
+                it.android.hardwareBack ?: true,
+                it.android.pauseMedia ?: true
+            )
+        }
+    }
+
 }